+++ assets/fixtures/fixture-3/Sample.jsx
@@ -1,0 +1,5 @@
+import styles from "./Sample.module.scss";
+
+export default function Sample() {
+  return <div className={styles.container}></div>;
+}